Backend/requirements.txt
Flask==3.0.1
Flask_Cors==4.0.0
Flask_PyMongo==2.3.0
pymongo==4.6.1
python-dotenv==1.0.1
